[refac] move SuccessCode.java, ErrorCode.java
