+++ src/main/java/org/hankki/hankkiserver/common/exception/UnauthorizedException.java
@@ -1,6 +1,6 @@
 package org.hankki.hankkiserver.common.exception;
 
-import org.hankki.hankkiserver.common.response.code.ErrorCode;
+import org.hankki.hankkiserver.common.code.ErrorCode;
 
 public class UnauthorizedException extends BusinessException {
     public UnauthorizedException() {
