+++ src/main/java/org/hankki/hankkiserver/common/response/BaseResponse.java
@@ -1,8 +1,8 @@
 package org.hankki.hankkiserver.common.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import org.hankki.hankkiserver.common.response.code.ErrorCode;
-import org.hankki.hankkiserver.common.response.code.SuccessCode;
+import org.hankki.hankkiserver.common.code.ErrorCode;
+import org.hankki.hankkiserver.common.code.SuccessCode;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
