+++ src/main/java/org/hankki/hankkiserver/external/openfeign/apple/ApplePublicKeyGenerator.java
@@ -1,6 +1,6 @@
 package org.hankki.hankkiserver.external.openfeign.apple;
 
-import org.hankki.hankkiserver.common.response.code.ErrorCode;
+import org.hankki.hankkiserver.common.code.ErrorCode;
 import org.hankki.hankkiserver.common.exception.UnauthorizedException;
 import org.hankki.hankkiserver.external.openfeign.apple.dto.ApplePublicKeys;
 import org.hankki.hankkiserver.external.openfeign.apple.dto.ApplePublicKey;
